remove all unnecessary bucket references

+++ backend/corpora/common/utils/aws.py
@@ -96,7 +96,8 @@ def delete_many_from_s3(bucket_name: str, dataset_uuid: str) -> None:
     This deletes everything with a specific prefix from the given bucket
 
     """
-    assert dataset_uuid
+    if not dataset_uuid:
+        raise ValueError
     s3 = boto3.resource("s3")
     bucket = s3.Bucket(bucket_name)
     bucket.objects.filter(Prefix=f"{dataset_uuid}/").delete()
